Move outer types into g.file
