+++ src/features/board/board.api.ts
@@ -1,23 +1,6 @@
 import { drelloApiAxios } from "src/utils/axios/drello-api-axios";
 import { path } from "src/utils/url/drello-api";
-
-export interface Board {
-  id: number;
-  title: string;
-  columns?: Column[];
-  cards?: Card[];
-}
-export interface Column {
-  id: number;
-  title: string;
-  boardId: number;
-}
-export interface Card {
-  id: number;
-  title: string;
-  columnId: number;
-  position: number;
-}
+import { OuterBoard } from "./board.g";
 
 export const getBoard = async ({
   boardId,
@@ -26,7 +9,7 @@ export const getBoard = async ({
   boardId: number;
   idToken: string;
 }) => {
-  return await drelloApiAxios.get<Board>(path.boards(boardId), {
+  return await drelloApiAxios.get<OuterBoard>(path.boards(boardId), {
     headers: { Authorization: `Bearer ${idToken}` },
   });
 };
