+++ src/features/card/card.g.ts
@@ -1,3 +1,10 @@
+export interface OuterCard {
+  id: number;
+  title: string;
+  columnId: number;
+  position: number;
+}
+
 export interface Card {
   id: number;
   title: string;
