+++ src/features/column/column.g.ts
@@ -1,3 +1,9 @@
+export interface OuterColumn {
+  id: number;
+  title: string;
+  boardId: number;
+}
+
 export interface Column {
   id: number;
   title?: string;
